Refactor horoscope logic to be grouped in one place, and adds cron job
