+++ src/pages/api/cron.ts
@@ -1,7 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import getHoroscope from "@/utils/horoscope";
-import { getTodaysDate } from "@/utils";
-import { getSupabase } from "@/utils/supabase";
+import processDailyHoroscope from "@/utils/horoscope";
 
 type Data = {
   success: boolean;
@@ -11,50 +9,17 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<Data>
 ) {
-  const supabase = getSupabase();
-  const formattedDate = getTodaysDate(new Date());
-  const checkExisting = await supabase
-    .from("horoscopes")
-    .select()
-    .eq("date", formattedDate);
+  const result = await processDailyHoroscope();
 
-  if (checkExisting?.data?.length) {
+  if (!result) {
     res.status(204).end();
     return;
   }
-
-  const zodiacSigns = [
-    "Capricorn",
-    "Aquarius",
-    "Pisces",
-    "Aries",
-    "Taurus",
-    "Gemini",
-    "Cancer",
-    "Leo",
-    "Virgo",
-    "Libra",
-    "Scorpio",
-    "Sagittarius",
-  ];
-
-  const horoscopes = await Promise.all(
-    zodiacSigns.map(async (sign) => {
-      const horoscope = await getHoroscope(sign);
-      return {
-        zodiac_sign: sign,
-        date: formattedDate,
-        description: horoscope,
-      };
-    })
-  );
-
-  const { data, error } = await supabase.from("horoscopes").insert(horoscopes);
-  if (error) {
-    console.error("Error inserting data into Supabase:", error);
+  if (result.error) {
+    console.error("Error inserting data into Supabase:", result.error);
     res.status(500).end();
   } else {
-    console.log("Data inserted into Supabase:", data);
+    console.log("Data inserted into Supabase:", result.data);
     res.status(200).json({ success: true });
   }
 }
